website/requirements.txt
altair==4.2.2
asgiref==3.6.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==39.0.1
cssselect==1.2.0
DateTime==5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.6
django-allauth==0.52.0
django-auto-logout==0.5.1
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-timezone-field==5.0
entrypoints==0.4
et-xmlfile==1.1.0
fanstatic==1.3
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.9.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
js.fullcalendar==2.9.1
js.jquery==3.6.1
js.momentjs==2.13.1
jsonschema==4.17.3
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
platformdirs==3.0.0
protobuf==3.20.0
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.6.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==13.3.4
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
sqlparse==0.4.3
streamlit==1.21.0
streamlit-lottie==0.0.3
tinysegmenter==0.3
tldextract==3.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.14
validators==0.20.0
virtualenv==20.19.0
watchdog==3.0.0
WebOb==1.8.7
zipp==3.15.0
zope.interface==6.0
